fix: '--continue' option
in addition to this commit,
the following dependency updates are required:
{
  "@warren-bank/node-denodeify": "^2.1.1"
}

+++ bin/nget/process_argv.js
@@ -221,8 +221,7 @@ if (argv_vals["--output-document"] && (argv_vals["--output-document"] !== "-")) 
       console.log('ERROR: Output file already exists')
       process.exit(0)
     }
-    else {
-      // "--continue" is not supported
+    if (!argv_vals["--continue"]) {
       fs.unlinkSync(argv_vals["--output-document"])
     }
   }
